xtuff/personal-time-management/integrate personal_finanial_assessment/pyproject.toml
[project]
name = "personal-financial-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "schedule>=1.2.2",
    "streamlit>=1.45.0",
]
